Drop krestful dependency and simplify things

+++ classes/aurora/controller/api.php
@@ -31,65 +31,37 @@
  */
 class Aurora_Controller_API extends Controller_REST
 {
-
-	protected $expire		 = true;
-	protected $authenticate	 = true;
-	protected $autorize		 = true;
-	/**
-	 * Used to construct the respective Model_(_common_name) & Collection_(_common_name)
-	 * by convention the controller should be called Controller_API_(_common_name)
-	 *
-	 * @var string
-	 */
-	protected $cname;
-	/**
-	 * Overrided krestful Restful_Controller
-	 * @var array
-	 */
-	protected $_accept		 = array(
-		'application/json'
-	);
-	protected $_accept_lang	 = array(
-		'en_US',
-		'en_GB'
-	);
-	protected $_accept_charset = array(
-		'utf-8',
-		'ISO-8859-1'
-	);
-	protected $_accept_strict	 = TRUE;
-	public function __construct(Request $request, Response $response, array $accept = NULL, array $accept_charset = NULL, array $accept_language = NULL, array $accept_strict = NULL) {
-		parent::__construct($request, $response, $accept, $accept_charset, $accept_language, $accept_strict);
-		$this->cname = Au::type()->cname($this);
-	}
 	public function action_index() {
-		$id		 = $this->request->param('id', NULL);
-		$m_or_c	 = Au::load($this->cname, $id);
+		$cname = Au::type()->cname($this);
+		$id = $this->request->param('id', NULL);
+		$m_or_c = Au::load($cname, $id);
 		$view = Au::json_encode($m_or_c);
 		$this->response->body($view->render());
 	}
 	public function action_create() {
-		$m = Au::json_decode($this->cname, $this->request->body());
+		$cname = Au::type()->cname($this);
+		$m = Au::json_decode($cname, $this->request->body());
 		Au::save($m);
 		$view = Au::json_encode($m);
 		$this->response->body($view);
 	}
 	public function action_update() {
-		$m = Au::json_decode($this->cname, $this->request->body());
+		$cname = Au::type()->cname($this);
+		$m = Au::json_decode($cname, $this->request->body());
 		Au::save($m);
 		$view = Au::json_encode($m);
 		$this->response->body($view);
 	}
 	public function action_delete() {
-		$m = Au::load($this->cname, $id);
+		$cname = Au::type()->cname($this);
+		$id = $this->request->param('id', NULL);
+		if (is_null($id))
+			throw new HTTP_Exception_404('Model does not exists');
+		$m = Au::load($cname, $id);
 		if (Au::is_new($m))
 			throw new HTTP_Exception_500('Model not loaded. Can not delete');
 		Au::delete($m);
 		$view = Au::json_encode($m);
 		$this->response->body($view);
-	}
-	protected function authorize() {
-		if (!$this->user->can($this->request->uri(), $this->request->action()))
-			throw new HTTP_Exception_403;
 	}
 }
